entity: importanti modifiche ai costruttori, specie riguardo le annotazioni @NonNull

+++ backend/src/main/java/com/dietiEstates/backend/model/entity/RealEstateForRent.java
@@ -3,6 +3,8 @@ package com.dietiEstates.backend.model.entity;
 
 import java.time.LocalDateTime;
 
+import org.modelmapper.internal.bytebuddy.implementation.bind.annotation.Super;
+
 import com.dietiEstates.backend.enums.EnergyClass;
 import com.dietiEstates.backend.model.embeddable.ExternalRealEstateFeatures;
 import com.dietiEstates.backend.model.embeddable.InternalRealEstateFeatures;
@@ -12,11 +14,13 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.ForeignKey;
 import jakarta.persistence.PrimaryKeyJoinColumn;
 import jakarta.persistence.Table;
-
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import lombok.experimental.SuperBuilder;
 
 
 
@@ -38,7 +42,7 @@ public class RealEstateForRent extends RealEstate
     @Column(name = "contract_years",
             nullable = false, 
             updatable = true)
-    private double contractYears;
+    private int contractYears;
